+++ Changes.rst
@@ -6,6 +6,9 @@ Unreleased Changes
 * Introduced `BodyFollowing` class for use with *body* parameter of
   `~HTTPConnection.send_request` method.
 
+* `~HTTPConnection.send_request` now returns a `HTTPResponse` instance
+  instead of a tuple.
+
 Release 1.0 (2013-07-13)
 ========================